rm load av_region module, replace functions
